% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflverse-package.R
\docType{package}
\name{nflverse-package}
\alias{nflverse}
\alias{nflverse-package}
\title{The 'nflverse' - A set of packages dedicated to data of the
National Football League.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'nflverse' is a set of packages dedicated to data of the National Football League. This package is designed to make it easy to install and load multiple 'nflverse' packages in a single step. Learn more about the 'nflverse' at \url{https://nflverse.nflverse.com/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nflverse.nflverse.com/}
  \item \url{https://github.com/nflverse/nflverse}
  \item Report bugs at \url{https://github.com/nflverse/nflverse/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Carl \email{mrcaseb@gmail.com}

Authors:
\itemize{
  \item Ben Baldwin \email{bbaldwin206@gmail.com}
  \item Lee Sharpe
  \item Tan Ho \email{tan@tanho.ca} (\href{https://orcid.org/0000-0001-8388-5155}{ORCID})
}

}
